# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AutoDealership Enterprise Platform - QA Training & API Specification",
    "version": "25.0.0-Master-Ultimate",
...
    "contact": {
      "name": "AutoDealership QA & Platform Engineering Team",
      "email": "qa-support@autodealer.vn"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
      "url": "https://training-car-sell-system.vercel.app",
      "description": "Vercel Production / Staging Environment"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authjs.session-token",
        "description": "NextAuth v5 session cookie. Automatically maintained by browser sessions upon logging in at /login."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JSON Web Token (JWT) passed in 'Authorization: Bearer <token>' header."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "1. Authentication & Security"
        ],
        "summary": "Sign In (REST API - Returns JWT Token & Sets Session Cookie)",
        "description": "**Premier RESTful Authentication for QA / API Testers:**\nAuthenticates user using Email or Phone with Password (or Sandbox OTP \\`888888\\`).\nReturns a signed JWT Bearer token in \\`data.token\\` AND sets HTTP-only session cookies (\\`authjs.session-token\\`).\nTesters can copy \\`data.token\\` and click **Authorize 🔒** in Swagger UI to authenticate all secured requests!",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "identity"
                ],
                "properties": {
                  "identity": {
                    "type": "string",
                    "example": "admin@autodealer.vn",
                    "description": "Email or Phone number"
                  },
                  "password": {
                    "type": "string",
                    "example": "Admin@123",
                    "description": "Account password (if mode == password)"
                  },
                  "otp": {
                    "type": "string",
                    "example": "888888",
                    "description": "6-digit OTP code (if mode == otp)"
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "password",
                      "otp"
                    ],
                    "default": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sign in successful. JWT token returned and cookies set.",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "message": "Đăng nhập thành công",
                    "token": "eyJhbGciOiJkaXIiLCJlbmMiOiJBMjU2R0NNIn0...",
                    "user": {
                      "id": "u-admin-uuid",
                      "email": "admin@autodealer.vn",
                      "fullName": "Nguyễn Văn Admin",
                      "phone": "0900000001",
                      "role": "ADMIN",
                      "showroomId": "sr-hn-cg-uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ERR_VALIDATION or ERR_UI_005 (Mã OTP không đúng)"
          },
          "401": {
            "description": "ERR_UI_003 (Mật khẩu sai) hoặc ERR_UI_004 (Tài khoản bị khóa 30p do nhập sai 5 lần)"
          }
        }
      }
    },
...
  }
}